added game model updated JwtAuthenticationFilter added GlobalExceptionHandler almost ready for post module

+++ src/main/java/com/zanta/lfp/repository/UserRepository.java
@@ -7,8 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User,Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
     boolean existsByUsername(String username);
+
     boolean existsByEmail(String email);
+
     Optional<User> findByUsername(String username);
 }
